requirements.txt
apache-airflow-client==2.2.0
dynaconf==3.1.7
fastapi==0.75.0
paramiko==2.10.3
pydantic==1.9.0
requests==2.27.1
SQLAlchemy==1.4.32
